requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
click==8.1.7
fastapi==0.115.2
h11==0.14.0
idna==3.10
pydantic==2.9.2
pydantic_core==2.23.4
sniffio==1.3.1
starlette==0.40.0
typing_extensions==4.12.2
uvicorn==0.32.0
pymongo==4.9.2
fastapi_clerk_auth==0.0.5
clerk_backend_api==1.2.0
openai-swarm==0.1.1
PyJWT>=2.0.0
python-dotenv==1.0.1
msgraph-core>=2.0.0
msal>=1.20.0
python-dotenv>=0.19.0